move authentication logic to separate store
